Refactored (#864)
