+++ client/src/featureform/providers/__init__.py
@@ -1,1 +1,2 @@
 from .providers import get_provider
+from .online_store import Scalar
