+++ client/src/featureform/providers/online_store.py
@@ -1,28 +1,85 @@
-from abc import ABC
+from abc import ABC, abstractmethod
+from typing import List, Any, Union
+from ..enums import ScalarType
+
+
+class ValueType(ABC):
+    @abstractmethod
+    def scalar(self):
+        pass
+
+    @abstractmethod
+    def is_vector(self):
+        pass
+
+
+class Scalar(ValueType):
+    def __init__(self, scalar_type: ScalarType):
+        self.scalar_type = scalar_type
+
+    def scalar(self):
+        return self.scalar_type
+
+    def is_vector(self):
+        return False
+
+
+class VectorType(ValueType):
+    def __init__(self, scalar_type: ScalarType, dimension: int, is_embedding: bool):
+        self.scalar_type = scalar_type
+        self.dimension = dimension
+        self.is_embedding = is_embedding
+
+    def scalar(self):
+        return self.scalar_type
+
+    def is_vector(self):
+        return True
 
 
 class OnlineStore(ABC):
-    def create_table(self, name, variant, entity_type):
+    @abstractmethod
+    def get_table(self, feature: str, variant: str):
         pass
 
-    def get_table(self, name, variant):
+    @abstractmethod
+    def create_table(self, feature: str, variant: str, entity_type: ValueType):
         pass
 
+    @abstractmethod
+    def delete_table(self, feature: str, variant: str):
+        pass
+
+    @abstractmethod
     def table_exists(self, name, variant):
         pass
 
-
-class OnlineTable(ABC):
-    def set(self, key, value):
-        pass
-
-    def get(self, key):
-        pass
-
+    @abstractmethod
     def close(self):
         pass
 
 
-class OnlineVectorTable(ABC):
-    def nearest(self, vector):
+class OnlineStoreTable(ABC):
+    @abstractmethod
+    def set(self, entity: str, value: object):
+        pass
+
+    @abstractmethod
+    def get(self, entity: str):
+        pass
+
+
+class VectorStoreTable(OnlineStoreTable, ABC):
+    @abstractmethod
+    def nearest(self, feature: str, variant: str, vector: List[float], k: int):
+        pass
+
+
+class VectorStore(OnlineStore, ABC):
+    @abstractmethod
+    def create_index(self, feature: str, variant: str, vector_type: VectorType):
+        pass
+
+    @abstractmethod
+    def delete_index(self, feature: str, variant: str):
         pass
